requirements.txt
pandas
numpy
pylive
# primesense - must be patched version
mido
pyglet
